requirements.txt
dash
dash-bootstrap-components
pandas
SQLAlchemy
psycopg2-binary
python-dotenv
gunicorn
python-dotenv
dash-deck
openai
pydeck
ipywidgets
dash-leaflet